Разобрался наконец-то с машинкой(выскакивала ошибка, помог дебаг) и военкоматом, заново его переписал, была проблема с pom и настраивал долго, не было родительского
