+++ Lesson8/src/main/java/militaryOffice/Person.java
@@ -1,3 +1,5 @@
+package militaryOffice;
+
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
@@ -5,6 +7,7 @@ import lombok.ToString;
 @Setter
 @Getter
 @ToString
+
 public class Person {
     private String name;
     private int age;
@@ -12,15 +15,11 @@ public class Person {
     private String country;
     private String city;
 
-    public Person(String country, String city) {
-        this.country = country;
-        this.city = city;
-    }
-
     public Person(String name, int age, Sex sex, String country, String city) {
-        this(country, city);
         this.name = name;
         this.age = age;
         this.sex = sex;
+        this.country = country;
+        this.city = city;
     }
 }
