+++ pom.xml
@@ -22,7 +22,6 @@
         <module>Lesson10</module>
         <module>Lesson11</module>
         <module>Lesson12</module>
-        <module>Lesson13</module>
     </modules>
 
     <properties>
